refactor: UserService 위치 변경
