+++ src/main/java/until/the/eternity/dcs/domain/user/fake/FakeUserService.java
@@ -1,6 +1,7 @@
 package until.the.eternity.dcs.domain.user.fake;
 
 import org.springframework.stereotype.Service;
+import until.the.eternity.dcs.domain.user.application.UserService;
 import until.the.eternity.dcs.domain.user.entity.UserSummary;
 
 // todo User 관련 실제 동작하는 서비스로 교체 후 Fake는 삭제
